renamed kit base classes and escaped important json returns
