+++ phoenixc2/server/kits/http_reverse/handler.py
@@ -1,4 +1,4 @@
-from ..base_handler import BaseHandler
+from ..handler_base import BaseHandler
 
 
 class Handler(BaseHandler):
